basic component division
